Crud entity

+++ src/Controller/TaskController.php
@@ -20,11 +20,25 @@ class TaskController extends AbstractController
     #[Route('/', name: 'task_list')]
     public function index(TaskService $taskService): Response
     {
-        return $this->render('Tasks/list.html.twig',['tasks'=> $taskService->getPaginatedTasks()]);
+        $user = $this->getUser();
+        if ($user === null) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
+        return $this->render('Tasks/list.html.twig', ['tasks' => $taskService->getPaginatedTasks()]);
+    }
+
+    #[Route('/finished', name: 'task_list_finished')]
+    public function finishedTaskList(TaskService $taskService): Response
+    {
+        $user = $this->getUser();
+        if ($user === null) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
+        return $this->render('Tasks/list.html.twig', ['tasks' => $taskService->getPaginatedTasks()]);
     }
 
     #[Route('/new', name: 'task_create', methods: ['GET', 'POST'])]
-    public function new(Request $request, TaskRepository $taskRepository,EntityManagerInterface $em, SluggerInterface $slugger): Response
+    public function new(Request $request, TaskRepository $taskRepository, EntityManagerInterface $em): Response
     {
 
         //I create my form for new task
@@ -37,8 +51,6 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($user === null) {
-            
-            $this->addFlash('danger', 'Veuillez vous connecter pour ajouter un task');
             return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
         }
         //I check if I have a form and that it is valid
@@ -66,11 +78,11 @@ class TaskController extends AbstractController
     }
 
     #[Route('/edition/{id}', name: 'task_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, TaskRepository $taskRepository,EntityManagerInterface $em): Response
+    public function edit(Request $request, TaskRepository $taskRepository, EntityManagerInterface $em): Response
     {
-        $task = new Task;
+        $task = new Task();
         $taskId = $request->attributes->get('id');
-        $task = $taskRepository->findTaskWithUserAndCategory($taskId);
+        $task = $taskRepository->findTaskWithUser($taskId);
 
 
         // $this->denyAccessUnlessGranted('task_EDIT', $task);
@@ -81,8 +93,6 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($user === null) {
-            
-            $this->addFlash('danger', 'Veuillez vous connecter pour modifier une tâche');
             return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
         }
         //I check if I have a form and that it is valid
@@ -91,7 +101,7 @@ class TaskController extends AbstractController
             $task->setUpdateAt($now);
             $task->setTitle(strtoupper($task->getTitle()));
             $task->setUser($user);
-            
+
             $em->persist($task);
             $em->flush();
 
@@ -109,24 +119,33 @@ class TaskController extends AbstractController
             'form' => $form->createView()
         ]);
     }
-    
-    #[Route('/{id}/toggle', name: 'task_toggle')]
+
+    #[Route('/toggle/{id}', name: 'task_toggle')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
-        $task->setIsDone(!$task->isIsDone());
-        $em->flush($task);
+        if ($this->getUser()->getId() === $task->getUser()->getId() || $this->getUser()->getRoles()[0] == "ROLE_ADMIN") {
+            $task->setIsDone(!$task->isIsDone());
+            $em->flush($task);
 
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+            $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+
+            return $this->redirectToRoute('task_list');
+        }
+        $this->addFlash('danger', sprintf('La tâche %s n\'a pas été marquée comme faite.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
 
     #[Route('/delete/{id}', name: 'task_delete')]
-    public function delete( Task $task, TaskRepository $taskRepository): Response
+    public function delete(Task $task, TaskRepository $taskRepository): Response
     {
         // $this->denyAccessUnlessGranted('TRICK_DELETE', $trick);
-        $taskRepository->remove($task, true);
-        $this->addFlash('success', sprintf('La tâche a bien eté supprimée'));
+
+        if ($this->getUser()->getId() === $task->getUser()->getId() || $this->getUser()->getRoles()[0] == "ROLE_ADMIN") {
+            $taskRepository->remove($task, true);
+            $this->addFlash('success', sprintf('La tâche a bien eté supprimée'));
+        }
+
         return $this->redirectToRoute('task_list', [], Response::HTTP_SEE_OTHER);
     }
 }
